usability changes for vector.h
added correctly commented and named (including params) forward
declarations for my vector library. Removed a duplicate function and
changed anywhere a unit vector was mentioned to be instead a
normalised vector to keep it more consistent with how this is usually
referred to in game development which is what this lib is targeted at.